Add semver auto-bump on pre-commit using cargo-workspaces

+++ models/src/speed.rs
@@ -1,6 +1,6 @@
 use bevy::prelude::Component;
 use derive_more::{Add, AddAssign, Display, From, Mul, MulAssign};
 
-// Speed in pixels per second
+// Speed in tiles per second (1 tile = 5ft)
 #[derive(Component, From, Add, Mul, MulAssign, AddAssign, Display, Debug, Default)]
-pub struct Speed(pub f32);
+pub struct Speed(pub u16);
